Included initial settings UI
Provides a simple UI if no arguments are passed to the software. Currently the
UI does not actually start a ping process, simply shows the user the settings
they can edit. It has everything ready for spawning ping plots via a `tkinter
TopLevel` widget, I just wanted to get these changes ready before moving on to
getting the live plotting working.

This adds an input value for `core.Core.ping`'s `timeout` parameter, detailed
in #83.

This commit corresponds with #82, laying down the groundwork to providing a UI.
It also provides further proof that the discussion in #79 is worth looking
into.