Properly displayed CLI arguments, made popup work
Makes an unnecessary call to ani.frame_seq which throws a type error. The code however will not work without this, as it does not start executing the plot animation.

When the user tries to create multiple popups, functionality breaks as the old plots try to animate in the new window.

Fixed UI to properly display CLI arguments.